+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/CameraMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 360302c5bfd2f49f4a63a848e07d7d4c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
