+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Animations/Girl_Running_Animation.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 720f1ab5f716944af9c1d82add5d0832
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
