+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Animations/Player_AnimationController.controller
@@ -10,8 +10,8 @@ AnimatorStateMachine:
   m_Name: Base Layer
   m_ChildStates:
   - serializedVersion: 1
-    m_State: {fileID: 3585031864394047180}
-    m_Position: {x: 300, y: 210, z: 0}
+    m_State: {fileID: 6255384378283174560}
+    m_Position: {x: 322, y: 277, z: 0}
   m_ChildStateMachines: []
   m_AnyStateTransitions: []
   m_EntryTransitions: []
@@ -21,7 +21,7 @@ AnimatorStateMachine:
   m_EntryPosition: {x: 50, y: 120, z: 0}
   m_ExitPosition: {x: 800, y: 120, z: 0}
   m_ParentStateMachinePosition: {x: 800, y: 20, z: 0}
-  m_DefaultState: {fileID: 3585031864394047180}
+  m_DefaultState: {fileID: 6255384378283174560}
 --- !u!91 &9100000
 AnimatorController:
   m_ObjectHideFlags: 0
@@ -44,14 +44,14 @@ AnimatorController:
     m_IKPass: 0
     m_SyncedLayerAffectsTiming: 0
     m_Controller: {fileID: 9100000}
---- !u!1102 &3585031864394047180
+--- !u!1102 &6255384378283174560
 AnimatorState:
   serializedVersion: 6
   m_ObjectHideFlags: 1
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: jammo_Running
+  m_Name: Girl_Running_Animation
   m_Speed: 1
   m_CycleOffset: 0
   m_Transitions: []
@@ -64,7 +64,7 @@ AnimatorState:
   m_MirrorParameterActive: 0
   m_CycleOffsetParameterActive: 0
   m_TimeParameterActive: 0
-  m_Motion: {fileID: 7400000, guid: 609b2aaaeb5eb429486611229a87a950, type: 2}
+  m_Motion: {fileID: 7400000, guid: 720f1ab5f716944af9c1d82add5d0832, type: 2}
   m_Tag: 
   m_SpeedParameter: 
   m_MirrorParameter:
