+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Materials/TileMaterial.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5084b478d42046649b37404706c9b15
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
