+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Mixamo_Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74188444ca5d84cad95408976ac10009
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
