+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Prefabs/Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7e1f8c900e25d4f1bb2769053445506e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
