+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Prefabs/Three_Tiles_Group.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 770978db2b98d4c7e812a54e9cab8b16
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
