+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Prefabs/Tile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 98deb6b0bdc184c55a672aae4b79af74
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
