+++ USE_CORE/Assets/_USE_Tasks/FruitRunner/Resources/Textures/Character_Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 036913dbd94f6455785780959358327a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
