--- USE_CORE/Assets/_USE_Tasks/SustainedAttention/ObjectManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8f1f39fdd1fe64fa9a8d9cd577d84609
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
